% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionFile.R
\name{suggest_cleaning}
\alias{suggest_cleaning}
\title{Suggest cleaning actions based on diagnostics}
\usage{
suggest_cleaning(
  df,
  na_threshold = 90,
  suggest_constant = TRUE,
  suggest_numeric_str = TRUE,
  suggest_impute = TRUE
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{na_threshold}{Percentage threshold of missing data before removal is suggested (default 90)}

\item{suggest_constant}{Whether to suggest dropping constant columns (default TRUE)}

\item{suggest_numeric_str}{Whether to suggest converting numeric-like strings (default TRUE)}

\item{suggest_impute}{Whether to suggest imputation for numeric columns (default TRUE)}
}
\value{
A tibble with suggested actions.
}
\description{
Suggest cleaning actions based on diagnostics
}
