% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionFile.R
\name{summarize_numeric}
\alias{summarize_numeric}
\title{Summarize numeric columns with stats and suggested fill}
\usage{
summarize_numeric(df)
}
\arguments{
\item{df}{A data.frame or tibble}
}
\value{
A tibble of summary stats and recommended fill method per numeric column
}
\description{
Summarize numeric columns with stats and suggested fill
}
\examples{
#Summarize numeric columns
df <- data.frame(
  x = c(10, 20, 30, NA),
  y = c(100, 200, 300, 400),
  z = c("a", "b", "c", "d"),
  stringsAsFactors = FALSE
)
summarize_numeric(df)
}
